+++ CHANGELOG.md
@@ -29,6 +29,9 @@ According to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) , the `Unr
 -   [#59](https://github.com/FC4E-WP5/fc4eosc-PIDMR-api/pull/59) - PIDMR-146 Add "endpoint" field to Provider entity with mode-specific endpoints.
 -   [#60](https://github.com/FC4E-WP5/fc4eosc-PIDMR-api/pull/60) - PIDMR-154 Add Property to Provider Entity to Track DOI Reliance.
 -   [#61](https://github.com/FC4E-WP5/fc4eosc-PIDMR-api/pull/61) - PIDMR-160 Implement Communication with Keycloak Admin for User Role Assignment.
+-   [#62](https://github.com/FC4E-WP5/fc4eosc-PIDMR-api/pull/62) - PIDMR-159 Create an Endpoint to Promote User to Any Role.
+-   [#63](https://github.com/FC4E-WP5/fc4eosc-PIDMR-api/pull/63) - PIDMR-161 Implement Admin Endpoint for Promotion Request Approval.
+
 
 
 ## 2.1.0 - 2024-04-29